[markup&styles] 게시글 작성 파일 경로 변경 및 form 수정 #43, #44
